Move cache pull to first step(s) of job
Even though we will not use the ccache until later in the
workflow, due to github's LRU eviction strategy when we
are near the github cache size limit we want to retrieve
the cache early before any real work starts for all jobs
so that the oldest caches are evicted in the order one
wants.  This is a subtle artifact, but was seen with a
specific case where the cache was under high usage and
the steps for cache retrieval did not occur until much
later for some jobs.