Force external and libs to be installed first for macOS
Due to the way that `make install` works on macOS
it requires that the libraries are installed before the
programs.  For a parallel make, this ends up failing
some of the time (timing is everything).  The
library dependency should be fixed in the Makefile
(and CmakeLists.txt), but for now, attempt to
provide a workaround for macOS (let's see if it works)..

This could probably also be fixed by eliminating the
parallel make option (at least for the install).